requirements.txt
fastapi==0.111.0
uvicorn==0.40.0
pydantic==2.12.5
pydantic-settings==2.12.0
email-validator==2.3.0
sqlalchemy==2.0.49
alembic==1.14.1
psycopg[binary]==3.2.13
redis==5.3.1
PyJWT==2.11.0
bcrypt==4.3.0
pytest==9.0.2
httpx==0.28.1
